requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
openai>=1.66.0
python-dotenv>=1.0.0
azure-identity>=1.19.0
